+++ plugin/buildSrc/src/main/kotlin/PluginDependencies.kt
@@ -1,5 +1,5 @@
 object PluginVersions {
-    val kotlin = "1.3.72"
+    val kotlin = "1.4.10"
     val ktlintPlugin = "9.3.0"
     val gradlePublishPlugin = "0.12.0"
     val androidPlugin = "3.6.3"