+++ samples/kotlin-multiplatform-common/src/main/kotlin/org/jlleitschuh/gradle/ktlint/sample/kotlin/multiplatform/Sample.kt
@@ -1,8 +1,0 @@
-package org.jlleitschuh.gradle.ktlint.sample.kotlin.multiplatform
-
-/**
- * Each platform must implement a Sample class matching this signature.
- */
-expect class Sample() {
-    fun doPlatformThing(): CharSequence
-}